fix: resolve clippy warnings on hermes-mission-control branch
- Remove unused assignment to degenerate_stream_detected
- Remove unnecessary 'as u64' cast in event logger
- Remove now-unused 'mut' on degenerate_stream_detected

+++ src/api/control.rs
@@ -6786,7 +6786,7 @@ fn spawn_control_session(
                         }
                     }
                     Err(broadcast::error::RecvError::Lagged(n)) => {
-                        lag_total = lag_total.saturating_add(n as u64);
+                        lag_total = lag_total.saturating_add(n);
                         tracing::warn!(
                             dropped = n,
                             lag_total = lag_total,
